backend/requirements.txt
Django==4.2.1
flake8==6.0.0
python-dotenv==1.0.0
pillow==9.5.0
djangorestframework==3.14.0
djoser==2.2.0
django-filter==23.2